refactor(cli): move function
