+++ packages/core/src/lib/upload.ts
@@ -2,11 +2,8 @@ import { uploadToPortal } from '@code-pushup/portal-client';
 import { PersistConfig, Report, UploadConfig } from '@code-pushup/models';
 import { getLatestCommit, loadReport } from '@code-pushup/utils';
 import { jsonToGql } from './implementation/json-to-gql';
-import {
-  GlobalOptions,
-  normalizePersistConfig,
-  normalizeUploadConfig,
-} from './types';
+import { normalizeUploadConfig } from './normalize';
+import { GlobalOptions, normalizePersistConfig } from './types';
 
 export type UploadOptions = { upload: Required<UploadConfig> } & {
   persist: Required<PersistConfig>;
